+++ pkg/storage/bucket/s3/bucket_client.go
@@ -36,6 +36,7 @@ func newS3Config(cfg Config) (s3.Config, error) {
 	return s3.Config{
 		Bucket:    cfg.BucketName,
 		Endpoint:  cfg.Endpoint,
+		Region:    cfg.Region,
 		AccessKey: cfg.AccessKeyID,
 		SecretKey: cfg.SecretAccessKey.Value,
 		Insecure:  cfg.Insecure,